Update the README to use the correct namespace
Note: At the moment, I have modified the instructions in the deps.edn
section of the README to use my fork of the project. Once the
clj-commons team cuts a release, I / they can change it back to the
proper mvn-version. At the moment, the README was anyway pointing to
an outdated version of `factual/durable-queue 0.1.5`